+++ rules/nette-code-quality/tests/Rector/ArrayDimFetch/AnnotateMagicalControlArrayAccessRector/Fixture/nested_isset_unset.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\NetteCodeQuality\Tests\Rector\ArrayDimFetch\ChangeControlArrayAccessToAnnotatedControlVariableRector\Fixture;
+namespace Rector\NetteCodeQuality\Tests\Rector\ArrayDimFetch\AnnotateMagicalControlArrayAccessRector\Fixture;
 
 use Nette\Application\UI\Presenter;
 use Nette\Application\UI\Form;
@@ -26,7 +26,7 @@ final class NestedIssetUnsedPresenter extends Presenter
 -----
 <?php
 
-namespace Rector\NetteCodeQuality\Tests\Rector\ArrayDimFetch\ChangeControlArrayAccessToAnnotatedControlVariableRector\Fixture;
+namespace Rector\NetteCodeQuality\Tests\Rector\ArrayDimFetch\AnnotateMagicalControlArrayAccessRector\Fixture;
 
 use Nette\Application\UI\Presenter;
 use Nette\Application\UI\Form;
