+++ rules/nette-code-quality/tests/Rector/ArrayDimFetch/AnnotateMagicalControlArrayAccessRector/Fixture/skip_already.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\NetteCodeQuality\Tests\Rector\ArrayDimFetch\ChangeControlArrayAccessToAnnotatedControlVariableRector\Fixture;
+namespace Rector\NetteCodeQuality\Tests\Rector\ArrayDimFetch\AnnotateMagicalControlArrayAccessRector\Fixture;
 
 use Nette\Application\UI\Presenter;
 use Nette\Application\UI\Form;
